Fix bugs on cora

+++ src/graph-server/GPU-Computation/cu_matrix.cuh
@@ -54,7 +54,7 @@ class CuMatrix : public Matrix {
     unsigned long long nnz;
     EdgeType *csrVal;
     unsigned *csrColInd;
-    unsigned long long *csrRowPtr;
+    unsigned *csrRowPtr;
 
 };
